Add narrow and contract integration tests
Narrow integration tests can run locally and only need a running mongodb
server on localhost:27017. The tests assert the integration of the code with
a local instance and test the complete application by mocking/stubbing
external services such as Google Cloud Storage and Auth0. The contract
integration tests assert that external services that are used in production
such as Google Cloud Storage, Auth0 and MongoDB Atlas all work with the interface
of the main application. Therefore, contract tests cannot run locally and require
access to the internet.