Add end2end test
This test tests the main application which has to be running on localhost:8000
and is connected to all external services (Auth0, Atlas, GCS). The test does
not use any test client to send requests but sends real requests with Python's
requests package. Therefore, no stubs or mocks are in use.